+++ server/src/main/java/org/opensearch/search/SearchService.java
@@ -1981,7 +1981,8 @@ public class SearchService extends AbstractLifecycleComponent implements IndexEv
                     bigArrays,
                     scriptService,
                     multiBucketConsumerService.create(),
-                    pipelineTree
+                    pipelineTree,
+                    multiBucketConsumerService.getBreaker()
                 );
             }
         };