+++ server/src/main/java/org/opensearch/search/aggregations/InternalAggregation.java
@@ -34,6 +34,7 @@ package org.opensearch.search.aggregations;
 import org.opensearch.common.annotation.PublicApi;
 import org.opensearch.common.util.BigArrays;
 import org.opensearch.core.common.Strings;
+import org.opensearch.core.common.breaker.CircuitBreaker;
 import org.opensearch.core.common.io.stream.NamedWriteable;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
@@ -97,6 +98,7 @@ public abstract class InternalAggregation implements Aggregation, NamedWriteable
         private final PipelineTree pipelineTreeRoot;
 
         private boolean isSliceLevel;
+        private final CircuitBreaker breaker;
         /**
          * Supplies the pipelines when the result of the reduce is serialized
          * to node versions that need pipeline aggregators to be serialized
@@ -112,13 +114,30 @@ public abstract class InternalAggregation implements Aggregation, NamedWriteable
             ScriptService scriptService,
             Supplier<PipelineTree> pipelineTreeForBwcSerialization
         ) {
-            return new ReduceContext(bigArrays, scriptService, (s) -> {}, null, pipelineTreeForBwcSerialization);
+            return new ReduceContext(bigArrays, scriptService, (s) -> {}, null, pipelineTreeForBwcSerialization, null);
         }
 
         /**
          * Build a {@linkplain ReduceContext} to perform the final reduction.
          * @param pipelineTreeRoot The root of tree of pipeline aggregations for this request
          */
+        public static ReduceContext forFinalReduction(
+            BigArrays bigArrays,
+            ScriptService scriptService,
+            IntConsumer multiBucketConsumer,
+            PipelineTree pipelineTreeRoot,
+            CircuitBreaker breaker
+        ) {
+            return new ReduceContext(
+                bigArrays,
+                scriptService,
+                multiBucketConsumer,
+                requireNonNull(pipelineTreeRoot, "prefer EMPTY to null"),
+                () -> pipelineTreeRoot,
+                breaker
+            );
+        }
+
         public static ReduceContext forFinalReduction(
             BigArrays bigArrays,
             ScriptService scriptService,
@@ -130,7 +149,8 @@ public abstract class InternalAggregation implements Aggregation, NamedWriteable
                 scriptService,
                 multiBucketConsumer,
                 requireNonNull(pipelineTreeRoot, "prefer EMPTY to null"),
-                () -> pipelineTreeRoot
+                () -> pipelineTreeRoot,
+                null
             );
         }
 
@@ -139,7 +159,8 @@ public abstract class InternalAggregation implements Aggregation, NamedWriteable
             ScriptService scriptService,
             IntConsumer multiBucketConsumer,
             PipelineTree pipelineTreeRoot,
-            Supplier<PipelineTree> pipelineTreeForBwcSerialization
+            Supplier<PipelineTree> pipelineTreeForBwcSerialization,
+            CircuitBreaker breaker
         ) {
             this.bigArrays = bigArrays;
             this.scriptService = scriptService;
@@ -147,6 +168,7 @@ public abstract class InternalAggregation implements Aggregation, NamedWriteable
             this.pipelineTreeRoot = pipelineTreeRoot;
             this.pipelineTreeForBwcSerialization = pipelineTreeForBwcSerialization;
             this.isSliceLevel = false;
+            this.breaker = breaker;
         }
 
         /**
@@ -210,6 +232,9 @@ public abstract class InternalAggregation implements Aggregation, NamedWriteable
             multiBucketConsumer.accept(size);
         }
 
+        public CircuitBreaker getBreaker() {
+            return breaker;
+        }
     }
 
     protected final String name;